Add Named Destinations to Catalog so that bookmarks and links can access them.
The named destinations require access to page nodes, so created Pages object
that is made using PagesFactory (which contains the page-related code from
Catalog).
