+++ src/UglyToad.PdfPig/Content/Catalog.cs
@@ -1,8 +1,7 @@
 ﻿namespace UglyToad.PdfPig.Content
 {
     using System;
-    using System.Collections.Generic;
-    using Core;
+    using Outline;
     using Tokens;
     using Util.JetBrains.Annotations;
 
@@ -12,90 +11,24 @@
     /// </summary>
     public class Catalog
     {
-        private readonly IReadOnlyDictionary<int, PageTreeNode> pagesByNumber;
-
         /// <summary>
         /// The catalog dictionary containing assorted information.
         /// </summary>
         [NotNull]
         public DictionaryToken CatalogDictionary { get; }
 
-        /// <summary>
-        /// Defines the page tree node which is the root of the pages tree for the document.
-        /// </summary>
-        [NotNull]
-        public DictionaryToken PagesDictionary { get; }
+        internal NamedDestinations NamedDestinations { get; }
 
-        /// <summary>
-        /// The page tree for this document containing all pages, page numbers and their dictionaries.
-        /// </summary>
-        public PageTreeNode PageTree { get; }
-
-        /// <summary>
-        /// Number of discovered pages.
-        /// </summary>
-        public int? NumberOfDiscoveredPages => pagesByNumber?.Count;
+        internal Pages Pages { get; }
 
         /// <summary>
         /// Create a new <see cref="CatalogDictionary"/>.
         /// </summary>
-        internal Catalog(DictionaryToken catalogDictionary, DictionaryToken pagesDictionary,
-            PageTreeNode pageTree)
+        internal Catalog(DictionaryToken catalogDictionary, Pages pages, NamedDestinations namedDestinations)
         {
             CatalogDictionary = catalogDictionary ?? throw new ArgumentNullException(nameof(catalogDictionary));
-            PagesDictionary = pagesDictionary ?? throw new ArgumentNullException(nameof(pagesDictionary));
-            PageTree = pageTree ?? throw new ArgumentNullException(nameof(pageTree));
-
-            if (!pageTree.IsRoot)
-            {
-                throw new ArgumentException("Page tree must be the root page tree node.", nameof(pageTree));
-            }
-
-            var byNumber = new Dictionary<int, PageTreeNode>();
-            PopulatePageByNumberDictionary(pageTree, byNumber);
-            pagesByNumber = byNumber;
-        }
-
-        private static void PopulatePageByNumberDictionary(PageTreeNode node, Dictionary<int, PageTreeNode> result)
-        {
-            if (node.IsPage)
-            {
-                if (!node.PageNumber.HasValue)
-                {
-                    throw new InvalidOperationException($"Node was page but did not have page number: {node}.");
-                }
-
-                result[node.PageNumber.Value] = node;
-                return;
-            }
-
-            foreach (var child in node.Children)
-            {
-                PopulatePageByNumberDictionary(child, result);
-            }
-        }
-
-        internal PageTreeNode GetPageNode(int pageNumber)
-        {
-            if (!pagesByNumber.TryGetValue(pageNumber, out var node))
-            {
-                throw new InvalidOperationException($"Could not find page node by number for: {pageNumber}.");
-            }
-
-            return node;
-        }
-
-        internal PageTreeNode GetPageByReference(IndirectReference reference)
-        {
-            foreach (var page in pagesByNumber)
-            {
-                if (page.Value.Reference.Equals(reference))
-                {
-                    return page.Value;
-                }
-            }
-
-            return null;
+            Pages = pages ?? throw new ArgumentNullException(nameof(pages));
+            NamedDestinations = namedDestinations;
         }
     }
 }
