+++ src/UglyToad.PdfPig/Parser/PdfDocumentFactory.cs
@@ -176,14 +176,18 @@
                 crossReferenceTable.Trailer,
                 parsingOptions.UseLenientParsing);
 
+            var pageFactory = new PageFactory(pdfScanner, resourceContainer, filterProvider,
+                new PageContentParser(new ReflectionGraphicsStateOperationFactory()));
+
             var catalog = CatalogFactory.Create(
                 rootReference,
                 rootDictionary,
                 pdfScanner,
+                pageFactory,
+                parsingOptions.Logger,
                 parsingOptions.UseLenientParsing);
 
-            var pageFactory = new PageFactory(pdfScanner, resourceContainer, filterProvider,
-                new PageContentParser(new ReflectionGraphicsStateOperationFactory()));
+            
 
             var acroFormFactory = new AcroFormFactory(pdfScanner, filterProvider, crossReferenceTable);
             var bookmarksProvider = new BookmarksProvider(parsingOptions.Logger, pdfScanner);
