+++ src/UglyToad.PdfPig/PdfDocument.cs
@@ -14,6 +14,8 @@
     using Tokenization.Scanner;
     using Tokens;
     using Outline;
+    using Outline.Destinations;
+    using System.Linq;
     using Util.JetBrains.Annotations;
 
     /// <inheritdoc />
@@ -75,13 +77,12 @@
         /// </summary>
         public bool IsEncrypted => encryptionDictionary != null;
 
-        internal PdfDocument(
-            IInputBytes inputBytes,
-            HeaderVersion version, 
+        internal PdfDocument(IInputBytes inputBytes,
+            HeaderVersion version,
             CrossReferenceTable crossReferenceTable,
             IPageFactory pageFactory,
             Catalog catalog,
-            DocumentInformation information, 
+            DocumentInformation information,
             EncryptionDictionary encryptionDictionary,
             IPdfTokenScanner pdfScanner,
             ILookupFilterProvider filterProvider,
@@ -98,7 +99,7 @@
             this.parsingOptions = parsingOptions;
 
             Information = information ?? throw new ArgumentNullException(nameof(information));
-            pages = new Pages(catalog, pageFactory, pdfScanner);
+            pages = catalog.Pages;
             Structure = new Structure(catalog, crossReferenceTable, pdfScanner);
             Advanced = new AdvancedPdfDocumentAccess(pdfScanner, filterProvider, catalog);
             documentForm = new Lazy<AcroForm>(() => acroFormFactory.GetAcroForm(catalog));
